requirements.txt
hg+https://bitbucket.org/wkornewald/django-nonrel
hg+https://bitbucket.org/wkornewald/django-dbindexer
hg+https://bitbucket.org/wkornewald/djangotoolbox
git+https://github.com/django-mongodb-engine/mongodb-engine
django-autoslug
pymongo
gunicorn

